requirements.txt
python-chess
stockfish
python-lichess
